Code Refactoring
Take "the needed include file and the defined marco" be in the front of `main.c`.

Change the for loop in `phonebooh_opt.c(append)` to while loop, because the condition in for loop is too long to study.

Resolve the warning "'CPU_time' defined but not used". The method is add `#ifdef DEBUG` to `phonebook_opt.h` and `phonebook_opt.c`.

Correct the for loop in the `main.c` line 105. First, get the condition i = 0 out of for loop. This way can reduce the branch in for loop. Second, correct the error in original file. Change `pHead->pNext` to `pHead`.

Add in appropriate commit.